requirements.txt
Brotli==1.0.9
click==7.1.2
cvxopt==1.2.5
cvxpy==1.1.7
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dnspython==2.1.0
ecos==2.0.7.post1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
osqp==0.6.2.post0
pandas==1.2.1
Pillow==8.1.0
plotly==4.14.3
pybind11==2.6.1
pymongo==3.11.2
pyparsing==2.4.7
pyportfolioopt==1.2.7
python-dateutil==2.8.1
pytz==2020.5
qdldl==0.1.5.post0
retrying==1.3.3
scipy==1.6.0
scs==2.1.2
six==1.15.0
Werkzeug==1.0.1
